backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
aiomysql==0.2.0
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# HTTP client
httpx==0.28.1
aiohttp==3.11.11

# Stripe
stripe==11.4.1

# AI/LLM
openai==1.58.1

# Data processing
pandas==2.2.3
numpy==2.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Utilities
python-dotenv==1.0.1
